Add template support.
Put caching flag into config file for debug. And put in simple caching.

Make config file export the %Config variable.

Looks like if the stat call is not successful, it doesn't just send 0/undef into the
function it sends nothing into the function which messes with the prototype (I could make
the argument optional instead, that may be better).